Split out getProductDisplayCost
