+++ client/state/products-list/selectors/get-product-display-cost.js
@@ -1,0 +1,23 @@
+/**
+ * Internal dependencies
+ */
+import { getProductBySlug } from './get-product-by-slug';
+
+import 'state/products-list/init';
+
+/**
+ * Returns the display price of the specified product.
+ *
+ * @param {object} state - global state tree
+ * @param {string} productSlug - internal product slug, eg 'jetpack_premium'
+ * @returns {?string} the display price formatted in the user's currency (eg 'A$29.00'), or null otherwise
+ */
+export function getProductDisplayCost( state, productSlug ) {
+	const product = getProductBySlug( state, productSlug );
+
+	if ( ! product ) {
+		return null;
+	}
+
+	return product.cost_display;
+}
